Extract `SignatureType` to common entity
