+++ src/domain/common/entities/signature-type.entity.ts
@@ -1,0 +1,6 @@
+export enum SignatureType {
+  ContractSignature = 'CONTRACT_SIGNATURE',
+  ApprovedHash = 'APPROVED_HASH',
+  Eoa = 'EOA',
+  EthSign = 'ETH_SIGN',
+}
