+++ src/domain/messages/entities/__tests__/message-confirmation.builder.ts
@@ -1,10 +1,8 @@
 import { faker } from '@faker-js/faker';
 import { Builder, IBuilder } from '@/__tests__/builder';
-import {
-  MessageConfirmation,
-  SignatureType,
-} from '@/domain/messages/entities/message-confirmation.entity';
+import { MessageConfirmation } from '@/domain/messages/entities/message-confirmation.entity';
 import { getAddress } from 'viem';
+import { SignatureType } from '@/domain/common/entities/signature-type.entity';
 
 export function messageConfirmationBuilder(): IBuilder<MessageConfirmation> {
   return new Builder<MessageConfirmation>()
