+++ src/domain/messages/entities/message-confirmation.entity.spec.ts
@@ -1,8 +1,6 @@
+import { SignatureType } from '@/domain/common/entities/signature-type.entity';
 import { messageConfirmationBuilder } from '@/domain/messages/entities/__tests__/message-confirmation.builder';
-import {
-  MessageConfirmationSchema,
-  SignatureType,
-} from '@/domain/messages/entities/message-confirmation.entity';
+import { MessageConfirmationSchema } from '@/domain/messages/entities/message-confirmation.entity';
 import { faker } from '@faker-js/faker';
 import { getAddress } from 'viem';
 import { ZodError } from 'zod';
